% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_locations}
\alias{covid_locations}
\title{covid_locations}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 215 rows and 3 columns.
}
\usage{
covid_locations
}
\description{
covid_locations
}
\keyword{datasets}
